forge fmt

+++ src/core/TTCConstants.sol
@@ -6,7 +6,7 @@ contract TTCConstants {
     string public constant TTC_SYMBOL = "TTC";
     uint256 public constant ONE = 10 ** 18;
 
-    uint public constant MIN_POW_BASE     = 1 wei;
-    uint public constant MAX_POW_BASE     = (2 * ONE) - 1 wei;
-    uint public constant POW_PRECISION    = ONE / 10**12;
+    uint256 public constant MIN_POW_BASE = 1 wei;
+    uint256 public constant MAX_POW_BASE = (2 * ONE) - 1 wei;
+    uint256 public constant POW_PRECISION = ONE / 10 ** 12;
 }
