+++ src/core/TTCMath.sol
@@ -4,42 +4,27 @@ pragma solidity ^0.8.20;
 import "./TTCConstants.sol";
 
 contract TTCMath is TTCConstants {
-    function toi(uint a)
-        internal pure 
-        returns (uint)
-    {
+    function toi(uint256 a) internal pure returns (uint256) {
         return a / ONE;
     }
 
-    function floor(uint a)
-        internal pure
-        returns (uint)
-    {
+    function floor(uint256 a) internal pure returns (uint256) {
         return toi(a) * ONE;
     }
 
-    function add(uint a, uint b)
-        internal pure
-        returns (uint)
-    {
-        uint c = a + b;
+    function add(uint256 a, uint256 b) internal pure returns (uint256) {
+        uint256 c = a + b;
         require(c >= a, "ERR_ADD_OVERFLOW");
         return c;
     }
 
-    function sub(uint a, uint b)
-        internal pure
-        returns (uint)
-    {
-        (uint c, bool flag) = subSign(a, b);
+    function sub(uint256 a, uint256 b) internal pure returns (uint256) {
+        (uint256 c, bool flag) = subSign(a, b);
         require(!flag, "ERR_SUB_UNDERFLOW");
         return c;
     }
 
-    function subSign(uint a, uint b)
-        internal pure
-        returns (uint, bool)
-    {
+    function subSign(uint256 a, uint256 b) internal pure returns (uint256, bool) {
         if (a >= b) {
             return (a - b, false);
         } else {
@@ -47,37 +32,28 @@ contract TTCMath is TTCConstants {
         }
     }
 
-    function mul(uint a, uint b)
-        internal pure
-        returns (uint)
-    {
-        uint c0 = a * b;
+    function mul(uint256 a, uint256 b) internal pure returns (uint256) {
+        uint256 c0 = a * b;
         require(a == 0 || c0 / a == b, "ERR_MUL_OVERFLOW");
-        uint c1 = c0 + (ONE / 2);
+        uint256 c1 = c0 + (ONE / 2);
         require(c1 >= c0, "ERR_MUL_OVERFLOW");
-        uint c2 = c1 / ONE;
+        uint256 c2 = c1 / ONE;
         return c2;
     }
 
-    function div(uint a, uint b)
-        internal pure
-        returns (uint)
-    {
+    function div(uint256 a, uint256 b) internal pure returns (uint256) {
         require(b != 0, "ERR_DIV_ZERO");
-        uint c0 = a * ONE;
+        uint256 c0 = a * ONE;
         require(a == 0 || c0 / a == ONE, "ERR_DIV_INTERNAL"); // mul overflow
-        uint c1 = c0 + (b / 2);
+        uint256 c1 = c0 + (b / 2);
         require(c1 >= c0, "ERR_DIV_INTERNAL"); //  add require
-        uint c2 = c1 / b;
+        uint256 c2 = c1 / b;
         return c2;
     }
 
     // DSMath.wpow
-    function powi(uint a, uint n)
-        internal pure
-        returns (uint)
-    {
-        uint z = n % 2 != 0 ? a : ONE;
+    function powi(uint256 a, uint256 n) internal pure returns (uint256) {
+        uint256 z = n % 2 != 0 ? a : ONE;
 
         for (n /= 2; n != 0; n /= 2) {
             a = mul(a, a);
@@ -92,45 +68,38 @@ contract TTCMath is TTCConstants {
     // Compute b^(e.w) by splitting it into (b^e)*(b^0.w).
     // Use `powi` for `b^e` and `powK` for k iterations
     // of approximation of b^0.w
-    function pow(uint base, uint exp)
-        internal pure
-        returns (uint)
-    {
+    function pow(uint256 base, uint256 exp) internal pure returns (uint256) {
         require(base >= MIN_POW_BASE, "ERR_POW_BASE_TOO_LOW");
         require(base <= MAX_POW_BASE, "ERR_POW_BASE_TOO_HIGH");
 
-        uint whole  = floor(exp);   
-        uint remain = sub(exp, whole);
+        uint256 whole = floor(exp);
+        uint256 remain = sub(exp, whole);
 
-        uint wholePow = powi(base, toi(whole));
+        uint256 wholePow = powi(base, toi(whole));
 
         if (remain == 0) {
             return wholePow;
         }
 
-        uint partialResult = powApprox(base, remain, POW_PRECISION);
+        uint256 partialResult = powApprox(base, remain, POW_PRECISION);
         return mul(wholePow, partialResult);
     }
 
-    function powApprox(uint base, uint exp, uint precision)
-        internal pure
-        returns (uint)
-    {
+    function powApprox(uint256 base, uint256 exp, uint256 precision) internal pure returns (uint256) {
         // term 0:
-        uint a     = exp;
-        (uint x, bool xneg)  = subSign(base, ONE);
-        uint term = ONE;
-        uint sum   = term;
+        uint256 a = exp;
+        (uint256 x, bool xneg) = subSign(base, ONE);
+        uint256 term = ONE;
+        uint256 sum = term;
         bool negative = false;
 
-
-        // term(k) = numer / denom 
+        // term(k) = numer / denom
         //         = (product(a - i - 1, i=1-->k) * x^k) / (k!)
         // each iteration, multiply previous term by (a-(k-1)) * x / k
         // continue until term is less than precision
-        for (uint i = 1; term >= precision; i++) {
-            uint bigK = i * ONE;
-            (uint c, bool cneg) = subSign(a, sub(bigK, ONE));
+        for (uint256 i = 1; term >= precision; i++) {
+            uint256 bigK = i * ONE;
+            (uint256 c, bool cneg) = subSign(a, sub(bigK, ONE));
             term = mul(term, mul(c, x));
             term = div(term, bigK);
             if (term == 0) break;
@@ -146,5 +115,4 @@ contract TTCMath is TTCConstants {
 
         return sum;
     }
-
 }
